PEP8 and pylint cleanup
